Refactor config and logging setup
