--- core/node/infra/constants.go
@@ -1,5 +1,0 @@
-package infra
-
-const (
-	CHAIN_ID_LOCALHOST = 31337
-)
